Making some progress with this packaging/importing issue. temporarily using Thonny to work on this until I can re format everything.
think the main problem is that I
keep changing naming conventions
and trying different methods for
getting imports to work. starting
to get a decent consistency and
things are starting to work here
and there. but still need to do some editing.

in the meantime. defined a main file
at the root of the package structure.
this is going to handle arg parsing
and delegating to the rest of
the program, basically-
calling subpackage.main() based on
what is required based on the input
type provided. the parsing works
minus -t and -e because those take
keyword dicts and I dont know how
to work that with arg parse.