Last time, we broke some stuff trying to add capabilities to Inline class, of detecting inheritance and packaging. Since we have seperate modules for dealing with that, repaired what we had fixed. Tests look good from inline everything works as expected.
moved both the argument declarations
and the parsing of args during runtime
to seperate modules (utils.options and
parsing.main).

root package.main() now just calls
parsing.main(), does some diagnostics
on the inline(s) parsed (which work 100/100 right now). this seems
unneeded, but in the future,
it will take the parsed product of
parsing.main and turn that into
class dict and classes in time.