Trying to decide whether users should be able to edit a class name for an class in the display table.
Would be hard to do since the data is stored in self attribute and dict
(as key) so must be immutable

work around is to tell users that
they must delete the class and go back
to make a new one. probably going
to go that way, but it seems to be
running away froim the fight., yet logical

still having very troubling problems
with the import system and packaging
really have no idea what i am doing wrong,

Went to stackexchange with this problem.
Hoping for some insight on
how to finally put it to bed for good.