Solved the immutable type issue with classdict.
now theres another strange, inexplicable happening-
packages and parents are switched
in class dict so that
packages = parents and parents = packages

temp solved it by switching their appearance in __str__()

not a good final solution at all.
program will be so bug happy if u dont
address this asap.

got edit working in the process, but formatting of the display table
is not great- very finnicky

also got the main function working
from the cmd line. We can now
generate classes with our
class generator at last. buit
there is an indentation problem in the
generated classes. We also
had to strip away all the extra
features to get class generation
to work (but it does!),

in the future, after class gen
str method is debugged. id like the
class generator fn in misc_functions
to accept a classdict as its argument