Completed validate_package function it seems.
Again packaging is getting
cumbersome- maybe make it a sub
package of parser and include different
modules for validating different things.

Still need to make sure that fns
in validation return 0 or 1 based on test
success or failure- and in some rare
cases- an updated version
of whatever is validated- though
it would be much neater to export
this responsiblity to the object instance.