Beefed up test_validation- Most of the prompt based test cases have been taken care of. Still have these problems with option and inline parsing, but the source of the issue has been identified and will be taken care of soon.
We will write unit tests for the essential
modules where possible- not classgenerator, regular file, etc..

in future
Make a module for building packages
and non-class-file-modules. so
we can move forward with the
packaging sctions when done
our testing of parsing and conventions
and have guranteed certainty that
they work as expected.