Made some philsophizing comments in test_validation. Considering whether the classdict is a worthwhile thing to have in this program. the biggest con to be said is that it makes looping and nesting of inlines much hairier. the proposed alternative is to make the entire inline into a list of positional args like this ``` [classA : [attr1, attr2], [method1] ... ] if a decision is made towards eliminating the class dict from the prog . then it must be considered what additional utility if any the classdict provides
also make sure to do this on an experimental branch as it will surely
introduce various breaking changes.

README.md - specified how to specify
multiple modules/files in a single package.

May be smart and make sense to extend this syntax change to inlines
for coherence-

classA : attr1, attr2 : method

if multiple classes then:

classA / classB : attr1/attr2, attrA/attrB.

However, I am not sold on that. We have put
a lot of work into testing and writing
functions that assume the original behaviour. Consider a way to
mitigate this syntax difference between
packaging and class inlines. Once the packaging syntax is fully fleshed
out, formal designed, implemented and
tested, the rest of this prog will
be a cake walk.