Spread the documentation across a few files, README, Inline_spec, example rc file and dev notes.
This will make the readme file more digestable. Since this is likely the
first thing a user reads before using
the program, it needs to be quick and
informative. The other files
can be looked to for more detail about
a specific topic,

README.md is almost finished, needs
proof-read, formatting, activated links

developer notes is all the extra notes
helpful comments, it likely wont be read at all, but needs revision/
updates, all the extra stuff from above (README)

Inline_specs is important and should
be very readable and easy to scan.

example rc file is empty for now,
going to be doing this and implementation of it last
since we need to finish the program before customizing it.

Next, we should make a custom class for replacing classdict. This will
be laborious but eventually helpful.