Tried to move the path injection statements 2 init Did not work. So expect some pesky import error issues in the future.
Hacked away at trying to get the main
generator to work with multi inlines again.
no luck. its failing validation because
it thinks the extension part of the inline
is the class name. cant blame it but a very
irritating bug. Time to revise the class designs
again for inline. Dont see why we cant make two
class names, and one has the extensions filtered out

The hard part is figuring out how to do this.
regexes arent working for me, but that seems like the
right answer to this.