Made a few BIG changes to this program and the mini language syntax. Detailed them in the documentation.
big time revised README.md and inline_specification.md
both are nearly finished and I want to be sure of that
(doubled over, proof read, read by another pair of eyes)

basicaly want the mini langauge syntax and all program
details to be perfectly foolproof (or as near as possible)
before moving ahead.

changed a bunch of stuff this week and
with the original synatx already hard coded,
it broke a lot including the main prog and slowed
progress big time.

Here are the big changes to the syntax and program
this session:
- forward slash used to seperate members of field family
instead of , (these are switched). This had to be done
because the comma in the extension was snagging
when we tried to parse it.

the program/software is now a general purpose
file generator, instead of a class generator.

the default behaivor is to generate classes,
but it can be used for packaging, modules, testing
suites, abcs, and all these things should be available
from the jump.

Added a load of todos to Inline spec and README.
big on the list- need to decide if we want to keep
the ability to provide a { dict } containing the args
for optional arguments.

We added the ability to collapse flags like
-te instead of -t -e. this will introduce breaking changes,
fix and unittest, after the user oriented documentation
is rocksolid.