Added Links to our now working index I want to look into how to write reference links without the dash-seperated-words. sort of ugly
added, some more examples and details.
Worked on the formating. Still have a lot
more formatting to go. and readme
might want to add or delete some sections
based on additions to the inline spec.

Readme should really just serve as
an operating manual for someone looking
to pick up the program and start using asap

Inline _spec is where we enumerate all the details.