Didnt really change anything in inline I was just experimenting with something.
Finished proofread of Inline_spec. its
good to go as far as I am concerned.
But you should get another pair of eyes on it.

Got up to the features section in README.md with our proofread.

Got the relative link to another file
working in markdown, but not
linking to a specific place in the file
(would be a real nice touch).