Making some big progress on getting inheritance and multiple file inlines working.
right now were held up trying to
Inheritance_builder to cope with the
fact that file identifiers may or may not
have an extension.

Also another paradigm shift with the parser. WE cant pass inlines into
the sub parse handlers (mulit,
inheritance, packaging) because
of the nuance of the extensions (again)

Instead, these files wil be generators
that take a string arg of the inlines return a list of inlines to be
generated. Hence, the validation
of both the unparsed string
and the individual inlines must take place before ANY parsing