Integrated extenion into Inline.
The extenion part can be accssed with
regular property syntax:
class.parents is class.extension.parents
under the hood.

Re added the cleanse function to strip adn split all items in the inline
except options.

speaking of options. These got a nifty
little set up- We have a dictionary now
that contains all possible switches as
keys and their values to False. Then
when switches are provided (regularly
or colllapsed) then the property with
that switches name switches to true.

extension, and in the future, inline, will have many default args
in their output/buiiltins (str and repr)
that will allow them to return the
correct value. using this, you could
tell inline to show its extension (or not)
show the default values of the extesion, etc. This should help with
getting this classs to cooperate with
the rest of the program.

TODO: continue enhancing Inline and any additional classes that need
to be made. On all, include documentation, unttesting, validation,
should be back up to speed quickly

Also fixed the startup batch file,
our inline mini language syntax
has changed since written.