getting re aqquainted with the src code. Feeling a bit better now that im running the test functions on the sub classes of inline and seeing that they work as intended.
scratch that about going off main branch to start from
scratch. However, you may want to going to drawing board
to consider what changes need to be made to finish this
in a brief manner. Then make issues for suggested major
improvements and feature enhancements, bug fixes etc.

we could go way into the weeds with this project and
id like to avoid that. in my mind, the things it needs
to be an acceptable production ready prototype is this:
- all three mode- working and tested
- full support for packaging, inheritance, extensions, modules
all fields and prepend arguments.
- .rc file, optional args, testing and exporting
- graceful crash if error gets thrown anywhere in runtime
(BIG ONE: something ive been meaning to, and the main
cause of me saying we need to go back to drawing board-
constructors and main functions etc need to be run in try
catch and have their own internal bug handling.