Finished implementing the first step of the program, which is to validate the name and path that the user provided, and with their permission, create a "project home folder" where the generator will create the requested files.
Tried moving on to debugging parsing but it looks
like we will need to start over here, its soo tightly
coupled with validation. another feature that I think,
can likely get the boot. just throw an error if the user
tries to generate something that is invalid according to
python syntax, other things like casing, we might want to
consider skipping.

At the very best, the coupling of validation and parsing needs
to be eliminated, so that you can easily toggle validation
on or off, and can parse an inline without a validation
error going off. These are two seperate steps and should
be implemented apart from each other.