troubleshooting. refactoring. This is the first time in a while i worked on this an felt good about it.
refactored the constructor of extension, which worked
but not great. We have massively simplified it, but
possibly too much, the whole parsing process is done
in about 3-5 lines of code (thank you map and lambda!)
but we no longer have a way of distinguishing parents
from packaging, which is obviously a problem.